backend/requirements.txt
# --- AI Core (Pinned for Stability) ---
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
numpy<2.0.0

# --- Whisper & Audio ---
openai-whisper==20231117
git+https://github.com/m-bain/whisperx.git
librosa>=0.9.0
soundfile>=0.12.1

# --- Backend & API ---
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.20
httpx>=0.28.1
python-dotenv>=1.0.0
supabase>=2.4.0

# --- Tools ---
yt-dlp>=2024.8.6
ffmpeg-python>=0.2.0
deep-translator>=1.11.4
langdetect>=1.0.9
sentencepiece>=0.1.99